Rename stores folder to store
Rename stores folder to store
Fix components path in specs
Fix `getDOMNode` warning on `CommentForm.jsx`
