+++ client/app/bundles/Comments/startup/ClientApp.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Provider } from 'react-redux';
 
-import createStore from '../stores/commentsStore';
+import createStore from '../store/commentsStore';
 import NonRouterCommentScreen from '../components/NonRouterCommentScreen';
 
 const App = props => {
